+++ examples/simple.clj
@@ -2,7 +2,7 @@
   (:require [tg-clj.core :as tg]
             [tg-clj-server.poll :as tg-poll]
             [tg-clj-server.utils :as u]
-            [tg-clj-server.defaults :as defaults]
+            [tg-clj-server.defaults.poll :as defaults]
 
             [clojure.tools.logging :as log]))