+++ examples/simple.clj
@@ -54,7 +54,7 @@
 
 (comment
   ;; Run this to start
-  (def f (future (main)))
+  (def stop (main))
 
   ;; Available commands:
   ;; - /hello - Posts a message after the command is input
@@ -62,4 +62,4 @@
   ;; All other messages are replied to
 
   ;; Run this to stop
-  (future-cancel f))
+  (stop))